[add]farmer/customer_review index

+++ db/migrate/20221123004248_create_farmer_reviews.rb
@@ -3,7 +3,7 @@ class CreateFarmerReviews < ActiveRecord::Migration[6.1]
     create_table :farmer_reviews do |t|
       t.integer :customer_id
       t.integer :job_offer_id
-      t.integer :farmer_star
+      t.float :farmer_star
       t.text :farmer_comment
       t.timestamps
     end
